move PCRBatch to sanger package
