+++ src/main/java/ca/gc/aafc/seqdb/api/entities/sanger/PcrBatch.java
@@ -1,4 +1,4 @@
-package ca.gc.aafc.seqdb.api.entities;
+package ca.gc.aafc.seqdb.api.entities.sanger;
 
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
@@ -29,6 +29,10 @@ import org.hibernate.annotations.Type;
 import org.hibernate.annotations.TypeDef;
 
 import ca.gc.aafc.dina.entity.DinaEntity;
+import ca.gc.aafc.seqdb.api.entities.ContainerType;
+import ca.gc.aafc.seqdb.api.entities.PcrPrimer;
+import ca.gc.aafc.seqdb.api.entities.Region;
+import ca.gc.aafc.seqdb.api.entities.ThermocyclerProfile;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -116,5 +120,9 @@ public class PcrBatch implements DinaEntity {
   @Type(type = "list-array")
   @Column(name = "attachment", columnDefinition = "uuid[]")
   private List<UUID> attachment = Collections.emptyList();
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "container_type_id")
+  private ContainerType containerType;
   
 }
