+++ src/main/java/ca/gc/aafc/seqdb/api/repository/PcrBatchRepository.java
@@ -5,8 +5,8 @@ import ca.gc.aafc.dina.repository.DinaRepository;
 import ca.gc.aafc.dina.repository.external.ExternalResourceProvider;
 import ca.gc.aafc.dina.security.DinaAuthorizationService;
 import ca.gc.aafc.dina.service.DinaService;
-import ca.gc.aafc.seqdb.api.dto.PcrBatchDto;
-import ca.gc.aafc.seqdb.api.entities.PcrBatch;
+import ca.gc.aafc.seqdb.api.dto.sanger.PcrBatchDto;
+import ca.gc.aafc.seqdb.api.entities.sanger.PcrBatch;
 import lombok.NonNull;
 import org.springframework.boot.info.BuildProperties;
 import org.springframework.stereotype.Repository;
