+++ src/main/java/ca/gc/aafc/seqdb/api/repository/VocabularyReadOnlyRepository.java
@@ -14,11 +14,11 @@ import java.util.stream.Collectors;
 import ca.gc.aafc.seqdb.api.dto.vocabularies.BaseVocabularyDto;
 import ca.gc.aafc.seqdb.api.entities.PcrPrimer;
 import ca.gc.aafc.seqdb.api.entities.ThermocyclerProfile;
+import ca.gc.aafc.seqdb.api.entities.sanger.PcrBatch;
 import ca.gc.aafc.seqdb.api.entities.Product;
 import ca.gc.aafc.seqdb.api.entities.Protocol;
 import ca.gc.aafc.seqdb.api.entities.Region;
 import ca.gc.aafc.seqdb.api.entities.MolecularSample;
-import ca.gc.aafc.seqdb.api.entities.PcrBatch;
 import io.crnk.core.exception.ResourceNotFoundException;
 import io.crnk.core.queryspec.QuerySpec;
 import io.crnk.core.repository.ReadOnlyResourceRepositoryBase;
