+++ src/main/java/ca/gc/aafc/seqdb/api/service/PcrBatchService.java
@@ -7,7 +7,7 @@ import org.springframework.validation.SmartValidator;
 
 import ca.gc.aafc.dina.jpa.BaseDAO;
 import ca.gc.aafc.dina.service.DefaultDinaService;
-import ca.gc.aafc.seqdb.api.entities.PcrBatch;
+import ca.gc.aafc.seqdb.api.entities.sanger.PcrBatch;
 import lombok.NonNull;
 
 @Service
