+++ src/main/resources/db/changelog/db.changelog-master.xml
@@ -21,4 +21,5 @@
   <include file="db/changelog/migrations/16-Add_group_to_LibraryPrepBatch.xml"/>
   <include file="db/changelog/migrations/17-Add_reactionDate_and_attachments_to_PcrBatch.xml"/>
   <include file="db/changelog/migrations/18-Add_new_PCRBatchItem.xml"/>
+  <include file="db/changelog/migrations/19-Add_ContainerType_to_PCRBatch.xml"/>
 </databaseChangeLog>
