+++ src/test/java/ca/gc/aafc/seqdb/api/repository/PcrBatchItemRepositoryIT.java
@@ -7,16 +7,14 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import javax.inject.Inject;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import ca.gc.aafc.seqdb.api.dto.MolecularSampleDto;
-import ca.gc.aafc.seqdb.api.dto.PcrBatchDto;
 import ca.gc.aafc.seqdb.api.dto.PcrBatchItemDto;
+import ca.gc.aafc.seqdb.api.dto.sanger.PcrBatchDto;
 import ca.gc.aafc.seqdb.api.testsupport.fixtures.MolecularSampleTestFixture;
 import ca.gc.aafc.seqdb.api.testsupport.fixtures.PcrBatchItemTestFixture;
 import ca.gc.aafc.seqdb.api.testsupport.fixtures.PcrBatchTestFixture;
-import ca.gc.aafc.seqdb.api.testsupport.fixtures.PcrReactionTestFixture;
 import io.crnk.core.exception.ResourceNotFoundException;
 import io.crnk.core.queryspec.QuerySpec;
